[file-manager] refactor: create folders for use case files
