+++ file-manager/src/application/use-case/album/index.ts
@@ -1,0 +1,2 @@
+export * from './add-new-album.use-case'
+export * from './get-albums.use-case'
