+++ file-manager/src/application/use-case/file/index.ts
@@ -1,0 +1,2 @@
+export * from './add-new-file.use-case'
+export * from './get-files-by-album-id.use-case'
