+++ file-manager/src/application/use-case/index.ts
@@ -1,5 +1,3 @@
-export * from './add-new-user.use-case'
-export * from './add-new-file.use-case'
-export * from './add-new-album.use-case'
-export * from './get-albums.use-case'
-export * from './get-files-by-album-id.use-case'
+export * from './album'
+export * from './file'
+export * from './user'
